Revert adapter pattern for `Registry`
This is essentially reverting [this PR](https://github.com/edisonywh/rocketman/pull/6).

The reason for this is because upon more consideration, it is apparent
that there's no real need for Registry itself to be adapter-based, since
there's nothing else to swap to.

The initial idea to persist to Redis will be about persisting the job
queue, not the producer<->consumer relationship (what Registry stores),
thus this reverts that.
